# backend/src/main/kotlin/com/thredge/backend/api/dto/ThreadDtos.kt
package com.thredge.backend.api.dto

import com.thredge.backend.support.ValidationMessages
import com.thredge.backend.support.validation.NotBlankIfPresent
import jakarta.validation.constraints.NotBlank
import java.time.Instant

data class ThreadSummary(
        val id: String,
        val title: String,
        val lastActivityAt: Instant,
        val categories: List<CategorySummary>,
        val pinned: Boolean,
        val isMarkdown: Boolean,
)

data class ThreadDetail(
        val id: String,
        val title: String,
        val body: String?,
        val createdAt: Instant,
        val lastActivityAt: Instant,
        val categories: List<CategorySummary>,
        val pinned: Boolean,
        val isMarkdown: Boolean,
        val entries: List<EntryDetail>,
)

data class ThreadFeedItem(
        val id: String,
        val title: String,
        val body: String?,
        val createdAt: Instant,
        val lastActivityAt: Instant,
        val categories: List<CategorySummary>,
        val pinned: Boolean,
        val entryCount: Int,
        val isMarkdown: Boolean,
)

data class ThreadCreateRequest(
        @field:NotBlank(message = ValidationMessages.BODY_REQUIRED) val body: String = "",
        val categoryNames: List<String> = emptyList(),
)

data class ThreadUpdateRequest(
        @field:NotBlankIfPresent(message = ValidationMessages.TITLE_REQUIRED)
        val title: String? = null,
        @field:NotBlankIfPresent(message = ValidationMessages.BODY_REQUIRED)
        val body: String? = null,
        val isMarkdown: Boolean? = null,
        val categoryNames: List<String>? = null,
)