# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "thredge API",
    "description": "thredge backend API (Kotlin/Spring).",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:28080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/entries/{id}": {
      "delete": {
        "tags": [
          "entry-controller"
        ],
        "operationId": "hideEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "entry-controller"
        ],
        "operationId": "updateEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntryDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/entries/{id}/restore": {
      "patch": {
        "tags": [
          "entry-controller"
        ],
        "operationId": "restoreEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntryDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/entries/{id}/move": {
      "patch": {
        "tags": [
          "entry-controller"
        ],
        "operationId": "moveEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryMoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntryDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/entries/{id}/move-to": {
      "patch": {
        "tags": [
          "entry-controller"
        ],
        "operationId": "moveEntryTo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryMoveTargetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntryDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/threads/hidden": {
      "get": {
        "tags": [
          "thread-controller"
        ],
        "operationId": "listHidden",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseThreadSummary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/entries/hidden": {
      "get": {
        "tags": [
          "entry-controller"
        ],
        "operationId": "listHidden_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseEntryDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/entries/hidden/search": {
      "get": {
        "tags": [
          "entry-controller"
        ],
        "operationId": "searchHiddenEntries",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseEntryDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategorySummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ThreadSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "lastActivityAt": {
            "type": "string",
            "format": "date-time"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategorySummary"
            }
          },
          "pinned": {
            "type": "boolean"
          },
          "isMarkdown": {
            "type": "boolean"
          }
        }
      },
...
      "EntryDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "isMarkdown": {
            "type": "boolean"
          },
          "parentEntryId": {
            "type": "string"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "threadId": {
            "type": "string"
          }
        }
      },
...
      "EntryUpdateRequest": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          }
        }
      },
      "EntryMoveRequest": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string",
            "enum": [
              "UP",
              "DOWN"
            ]
          }
        },
        "required": [
          "direction"
        ]
      },
      "EntryMoveTargetRequest": {
        "type": "object",
        "properties": {
          "targetEntryId": {
            "type": "string",
            "minLength": 1
          },
          "position": {
            "type": "string",
            "enum": [
              "BEFORE",
              "AFTER",
              "CHILD"
            ]
          }
        },
        "required": [
          "position",
          "targetEntryId"
        ]
      },
      "PageResponseThreadSummary": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThreadSummary"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
...
      "PageResponseEntryDetail": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntryDetail"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}